+++ packages/xforms-engine/src/lib/instance-defaults.ts
@@ -1,0 +1,43 @@
+import { OPENROSA_XFORMS_PREFIX } from '@getodk/common/constants/xmlns';
+import type { InstanceValueContext } from '../instance/internal-api/InstanceValueContext';
+import type { AttributeContext } from '../instance/internal-api/AttributeContext';
+import type { InstanceDefaults } from '../client';
+
+type ValueContext = AttributeContext | InstanceValueContext;
+
+const PROTECTED_META_FIELDS = [
+  'instanceID',
+  'instanceName',
+  'timeStart',
+  'timeEnd',
+  'today',
+  'userID',
+  'deviceID',
+  'deprecatedID',
+  'email',
+  'phoneNumber',
+  'audit',
+];
+
+const isProtectedMetaProperty = (refWithoutRoot: string): boolean => {
+  return !!PROTECTED_META_FIELDS.find((field) => {
+    const withoutNamespace = refWithoutRoot.replaceAll(`${OPENROSA_XFORMS_PREFIX}:`, '');
+    return withoutNamespace === `meta/${field}`;
+  });
+};
+
+export const getInstanceDefaultValue = (
+  defaults: InstanceDefaults,
+  context: ValueContext
+): string | undefined => {
+  if (Object.keys(defaults).length === 0 || context.contextNode.nodeType === 'attribute') {
+    return;
+  }
+  const ref = context.contextReference();
+  const positionOfSecondSlash = ref.indexOf('/', 1);
+  const refWithoutRoot = ref.slice(positionOfSecondSlash + 1);
+  if (isProtectedMetaProperty(refWithoutRoot)) {
+    return;
+  }
+  return defaults[ref] ?? (refWithoutRoot && defaults[refWithoutRoot]) ?? undefined;
+};